Moved The responses into a responses folder & added a request folder
