+++ src/main/java/edu/fontysmaua/tournamentapi/business/TeamUseCases.java
@@ -1,15 +1,12 @@
 package edu.fontysmaua.tournamentapi.business;
 
-import edu.fontysmaua.tournamentapi.domain.Team.CreateTeamResponse;
-import edu.fontysmaua.tournamentapi.domain.Team.DeleteTeamResponse;
-import edu.fontysmaua.tournamentapi.domain.Team.GetAllTeamsResponse;
-import edu.fontysmaua.tournamentapi.domain.Team.UpdateTeamResponse;
-import edu.fontysmaua.tournamentapi.domain.Team.DeleteTeamResponse;
+import edu.fontysmaua.tournamentapi.domain.Team.response.CreateTeamResponse;
+import edu.fontysmaua.tournamentapi.domain.Team.response.DeleteTeamResponse;
+import edu.fontysmaua.tournamentapi.domain.Team.response.GetAllTeamsResponse;
+import edu.fontysmaua.tournamentapi.domain.Team.response.UpdateTeamResponse;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import com.mysql.cj.x.protobuf.MysqlxCrud.Delete;
 
 import edu.fontysmaua.tournamentapi.persistence.TeamRepository;
 import edu.fontysmaua.tournamentapi.persistence.entity.TeamEntity;
