+++ src/main/java/edu/fontysmaua/tournamentapi/business/impl/Team/CreateTeamImpl.java
@@ -4,7 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import edu.fontysmaua.tournamentapi.business.TeamUseCases;
-import edu.fontysmaua.tournamentapi.domain.Team.CreateTeamResponse;
+import edu.fontysmaua.tournamentapi.domain.Team.response.CreateTeamResponse;
 import edu.fontysmaua.tournamentapi.persistence.TeamRepository;
 import edu.fontysmaua.tournamentapi.persistence.entity.TeamEntity;
