+++ src/main/java/edu/fontysmaua/tournamentapi/controller/TeamsController.java
@@ -13,10 +13,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import edu.fontysmaua.tournamentapi.business.TeamUseCases;
-import edu.fontysmaua.tournamentapi.domain.Team.CreateTeamResponse;
-import edu.fontysmaua.tournamentapi.domain.Team.DeleteTeamResponse;
-import edu.fontysmaua.tournamentapi.domain.Team.GetAllTeamsResponse;
-import edu.fontysmaua.tournamentapi.domain.Team.UpdateTeamResponse;
+import edu.fontysmaua.tournamentapi.domain.Team.response.CreateTeamResponse;
+import edu.fontysmaua.tournamentapi.domain.Team.response.DeleteTeamResponse;
+import edu.fontysmaua.tournamentapi.domain.Team.response.GetAllTeamsResponse;
+import edu.fontysmaua.tournamentapi.domain.Team.response.UpdateTeamResponse;
 import edu.fontysmaua.tournamentapi.persistence.entity.TeamEntity;
 import lombok.RequiredArgsConstructor;
