+++ src/main/java/edu/fontysmaua/tournamentapi/domain/Team/request/SaveTeamRequest.java
@@ -1,0 +1,4 @@
+package edu.fontysmaua.tournamentapi.domain.Team.request;
+
+public class SaveTeamRequest {
+}
